+++ packages/card/src/index.style.tsx
@@ -1,37 +1,19 @@
 import styled from 'styled-components';
-import { bg, color, fg } from '@lendi-ui/color';
-import { p, my, px, ml, mr } from '@lendi-ui/spacing';
-import { Heading } from '@lendi-ui/typography';
-import { map, BreakpointValue, BreakpointValueMap } from '@lendi-ui/breakpoint';
-import { deriveSize } from '@lendi-ui/utils';
+import { bg, color } from '@lendi-ui/color';
+import { p, my, px, mr } from '@lendi-ui/spacing';
+import { Heading, body } from '@lendi-ui/typography';
 
-export type Size = 'lg' | 'sm';
-
-const headingStyleBySizeMixin = (size: BreakpointValue<Size> | BreakpointValueMap<Size>) =>
-  map(size, (val) => {
-    switch (val) {
-      case 'sm':
-        return `
-          font-size: ${deriveSize(1)};
-      `;
-      case 'lg':
-        return `
-          font-size: ${deriveSize(1.5)};
-      `;
-      default:
-        return `font-size: ${deriveSize(1)}`;
-    }
-  });
+export type Size = 'xs' | 'sm' | 'md';
 
 export const CardWrapper = styled.div`
   border: 1px solid ${color('shade.100')};
   position: relative;
-  border-radius: 12px;
+  border-radius: 6px;
   ${my('md')};
 `;
 
-export const CardHead = styled(Heading)`
-  border-radius: 12px 12px 0 0;
+export const CardHead = styled.div`
+  border-radius: 6px 6px 0 0;
   ${bg('shade.50')};
   min-height: 40px;
   margin-bottom: -1px;
@@ -41,7 +23,6 @@ export const CardHead = styled(Heading)`
   display: flex;
   align-content: center;
   justify-content: space-between;
-  ${({ size }: { size: BreakpointValue<Size> | BreakpointValueMap<Size> }) => headingStyleBySizeMixin(size)};
 `;
 
 export const TitleContainer = styled.div`
@@ -54,7 +35,7 @@ export const CardTitleIcon = styled.span`
   ${mr('xxxs')};
 `;
 
-export const CardTitle = styled.div`
+export const CardTitle = styled(Heading)`
   overflow-wrap: normal;
   word-wrap: break-word;
   flex: 1;
@@ -64,29 +45,18 @@ export const CardTitle = styled.div`
 export const CardHeadSubTitle = styled.div`
   overflow-wrap: normal;
   font-size: 12px;
+  ${body({ size: 'xs' })};
 `;
 
 export const MoreContainer = styled.div`
   display: flex;
   justify-content: space-evenly;
   align-items: center;
-  ${p('xxxs')};
+  ${p('xs')};
 `;
 
 export const CardBody = styled.div`
   ${p('lg')};
   background: transparent;
-`;
-
-export const CancelButton = styled.div`
-  border: none;
-  ${fg('primary.500')};
-  font-weight: bold;
-  cursor: pointer;
-`;
-
-export const MoreIcon = styled.div`
-  ${ml('lg')};
-  cursor: pointer;
-  ${p('xxxs')};
+  ${body({ size: 'xs' })};
 `;