+++ packages/card/src/index.test.tsx
@@ -17,19 +17,12 @@ const render = (props: CardProps) => {
 };
 
 describe('Card', () => {
-  it('renders card without Cancel button', () => {
+  it('renders card', () => {
     render({ title: 'My Testing card' });
     expect(wrapper.find(Card)).toHaveLength(1);
     expect(wrapper.find(CardHead)).toHaveLength(1);
     expect(wrapper.find(CardBody)).toHaveLength(1);
     expect(wrapper.html().indexOf('My Testing card')).not.toEqual(-1);
-    expect(wrapper.props().onCancel).toBeUndefined();
-  });
-
-  it('renders card with Cancel button', () => {
-    render({ title: 'My Testing card', onCancel: () => alert('on cancel clicked') });
-    expect(wrapper.find(MoreContainer)).toHaveLength(1);
-    expect(wrapper.find(Card).props().onCancel).toBeDefined();
   });
 
   it('render card with interactive title', () => {
@@ -48,41 +41,20 @@ describe('Card', () => {
     expect(wrapper.find(Card)).toMatchSnapshot();
   });
 
-  it('render card with icon and cancel button', () => {
+  it('render card with sm headersize', () => {
     render({
       title: 'My Testing card',
-      cardTitleIcon: <Lock color="secondary.500" />,
-      onCancel: () => alert('on cancel clicked'),
+      headerSize: 'sm',
     });
-    expect(wrapper.find(CardTitleIcon)).toHaveLength(1);
-    expect(wrapper.find(Card).props().cardTitleIcon).toBeDefined();
-    expect(wrapper.find(CardHead)).toHaveStyleRule('font-size', deriveSize(1)); // default size
+    expect(wrapper.find(Card).props().headerSize).toEqual('sm');
     expect(wrapper.find(Card)).toMatchSnapshot();
   });
-
-  it('render card with lg headersize', () => {
+  it('render card with rightHeaderContent', () => {
     render({
       title: 'My Testing card',
-      headerSize: 'lg',
-      onCancel: () => alert('on cancel clicked'),
+      rightHeaderContent: <div>hello world!</div>,
     });
-    expect(wrapper.find(Card).props().headerSize).toEqual('lg');
-    expect(wrapper.find(CardHead)).toHaveStyleRule('font-size', deriveSize(1.5));
-    expect(wrapper.find(Card)).toMatchSnapshot();
-  });
-  it('render card with just more icon', () => {
-    render({
-      title: 'My Testing card',
-      moreIcon: <Info color="secondary.500" />,
-      onIconClick: () => alert('Icon clicked!'),
-    });
-    expect(wrapper.find(Card).props().title).toEqual('My Testing card');
-    expect(wrapper.find(Card).props().moreIcon).toEqual(
-      expect.objectContaining({
-        props: { color: 'secondary.500' },
-      })
-    );
-    expect(wrapper.find(Card).props().onIconClick).toBeInstanceOf(Function);
+    expect(wrapper.find(Card).props().rightHeaderContent).toBeDefined();
     expect(wrapper.find(Card)).toMatchSnapshot();
   });
 });